new_stats: verbose mode, ci arg and pep8 cleanups
Mimic the behavior of do_stats in that it will only print statistically
significant results unless you specify a -v

To make that a bit more useful, also add an argument for the confidence
interval. It can be any percentage value written as a whole number, ie. 95 for
95% confidence interval

Address some of the pep8 garbage while here also