Add mann-whitney u variable and overrides
new_stats -tv would be the equivalent of the old new_stats -v

Otherwise, by default, the script follows what's in the comments. If the data is
normally distributed, the student t-test is used. Otherwise, mann-whitney.

NOTE: This is the first commit where I am willingly diverging with ministat.
AFAICT, ministat does not ever use mann-whitney.