Use nir_validate_shader to determine debug build
The old method looked for a section called debug. This meant if you used -g,
even with something like -O3, the check would fail. This new method is a little
more fragile in that if nir_validate_shader ever evaluates to anything in a perf
build, we need to try again. Also, it doesn't exist pre-nir.

meh

Recommended-by: Kenneth Graunke <kenneth@whitecape.org>