Move tests into separate files (except synmark)
To keep the main runner clean,, and provide the ability to easily plug in new
benchmarks, move test command initialization to separate files.

The existing synmark config was a bit convoluted. I need to fix that before I
push it into a separate file.

NOTE: I really didn't want to do this. I liked the idea of having a simple array
list of tests in an obvious place. However, aside from the fact that things grew
out of hand, there is a need for people to be able to add their own stuff.